feat(cmd/influx): modify the query cli to use the http api
This modifies the query cli to remove the flux runtime dependency and
uses the public go client to execute the query. It then uses flux's csv
decoder to read the raw response.

This removes any dependency on the flux runtime and substantially
reduces the binary size and the startup execution time.